requirements.txt
langchain
pypdf2
qdrant-client
openai
flask
PyYAML
